✨feat: 워치와 연동한 데이터 활용

+++ lib/model/record_model.dart
@@ -5,7 +5,7 @@ class Record {
   final String? exerciseName;
   final DateTime? exerciseDate;
   final int? successCount;
-  final int? caloriesBurnedSum;
+  final double? caloriesBurnedSum;
   final int? averageHeartRate;
 
   Record(
